PSI_1_010_harshal_plant-disease-identifier/Backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
sqlalchemy
pillow>=10.0.0
python-multipart
gunicorn